+++ NEWS.md
@@ -3,4 +3,5 @@ ClimaUtilities.jl Release Notes
 
 v0.1.3
 -------
+- Add `DataStructures` module containing `LRUCache` object. PR [#35](https://github.com/CliMA/ClimaUtilities.jl/pull/35)
 - Add `OutputPathGenerator`. PR [#28](https://github.com/CliMA/ClimaLand.jl/pull/28)